+++ src/Nethermind/Nethermind.Blockchain.Test/Validators/BlockValidatorTests.cs
@@ -15,13 +15,14 @@
 //  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
 
 using Nethermind.Blockchain.Validators;
+using Nethermind.Core;
 using Nethermind.Core.Specs;
-using Nethermind.Specs;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Logging;
+using Nethermind.Specs;
 using NUnit.Framework;
 
-namespace Nethermind.Core.Test
+namespace Nethermind.Blockchain.Test.Validators
 {
     [TestFixture]
     public class BlockValidatorTests
